reorganizing test packages and making library more hermetic
