+++ library/src/test/java/com/github/pwittchen/reactivenetwork/library/network/observing/strategy/MarshmallowNetworkObservingStrategyTest.java
@@ -1,0 +1,86 @@
+/*
+ * Copyright (C) 2016 Piotr Wittchen
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.github.pwittchen.reactivenetwork.library.network.observing.strategy;
+
+import android.app.Application;
+import android.net.NetworkInfo;
+import com.github.pwittchen.reactivenetwork.library.BuildConfig;
+import com.github.pwittchen.reactivenetwork.library.Connectivity;
+import com.github.pwittchen.reactivenetwork.library.network.observing.NetworkObservingStrategy;
+import org.junit.Rule;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.mockito.Spy;
+import org.mockito.junit.MockitoJUnit;
+import org.mockito.junit.MockitoRule;
+import org.robolectric.RobolectricTestRunner;
+import org.robolectric.RuntimeEnvironment;
+import org.robolectric.annotation.Config;
+import rx.Observable;
+import rx.Subscription;
+import rx.functions.Action1;
+
+import static com.google.common.truth.Truth.assertThat;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
+
+@RunWith(RobolectricTestRunner.class) @Config(constants = BuildConfig.class)
+public class MarshmallowNetworkObservingStrategyTest {
+
+  @Rule public MockitoRule rule = MockitoJUnit.rule();
+  @Spy private NetworkObservingStrategy strategy = new MarshmallowNetworkObservingStrategy();
+
+  @Test public void shouldObserveConnectivity() {
+    // given
+    final NetworkObservingStrategy strategy = new MarshmallowNetworkObservingStrategy();
+
+    // when
+    strategy.observeNetworkConnectivity(RuntimeEnvironment.application)
+        .subscribe(new Action1<Connectivity>() {
+          @Override public void call(Connectivity connectivity) {
+
+            // then
+            assertThat(connectivity.getState()).isEqualTo(NetworkInfo.State.CONNECTED);
+          }
+        });
+  }
+
+  @Test public void shouldStopObservingConnectivity() {
+    // given
+    final NetworkObservingStrategy strategy = new MarshmallowNetworkObservingStrategy();
+    final Application context = RuntimeEnvironment.application;
+    final Observable<Connectivity> observable = strategy.observeNetworkConnectivity(context);
+
+    // when
+    final Subscription subscription = observable.subscribe();
+    subscription.unsubscribe();
+
+    // then
+    assertThat(subscription.isUnsubscribed()).isTrue();
+  }
+
+  @Test public void shouldCallOnError() {
+    // given
+    final String message = "error message";
+    final Exception exception = new Exception();
+
+    // when
+    strategy.onError(message, exception);
+
+    // then
+    verify(strategy, times(1)).onError(message, exception);
+  }
+}
